Move edits to correct class
