+++ GeoTagging/app/src/main/java/com/mobilapp/geotagging/Tag.java
@@ -26,4 +26,5 @@ public class Tag {
     /* set latitude with myLocation.setLongitude(latitude) */
     @ColumnInfo(name = "latitude")
     public double latitude;
+
 }
