+++ GeoTagging/app/src/main/java/com/mobilapp/geotagging/TagDao.java
@@ -17,4 +17,5 @@ public interface TagDao {
 
     @Insert
     public void insertNewTag(Tag tag);
+
 }
